First functional... something.
Requires AirBreather.Common.  See my other repository.

7za-*.exe: command-line 7-zip excutable.
CachedMd5: some static helpers to calculate MD5 hashes or reuse if there's already a cached version from a previous run.
Extensions: Task.Finally does a bit of a try / finally pattern for tasks, and StringBuilder.MoveToString is a combined ToString + Clear.
Options: Standard CommandLineParser stuff.
ProcessRunner: helper to stitch various arguments back together and run a process asynchronously, TAP-style.
SetupTasks: the "doers" of this.  Currently just ExtractArchive (see what it says on the tin for what it does) and a placeholder for Clean.
SevenZipExtractor: layer on top of ProcessRunner and the packed 7za-*.exe resource to extract any archive that 7-zip is able to extract.
Program: indexes the files in the input directories by MD5 checksum, collates them with the universe of files needed for the setup process, downloads missing ones that can be downloaded, reports on the missing ones that could not be downloaded (or whose checksums failed post-download), and then runs all the setup tasks.
STEP Core 2.2.9.2.xml: machine-readable definition of STEP Core 2.2.9.2, for processing with this tool.