It begins: BSA files, the semi-hard way.
I listed a few contingencies in #12 that I said I could try before rolling it myself.  Here's why I chose none of them:

- BSAopt does some very evil things with its I/O that makes it impossible to automate nicely and difficult to automate at all.  It steals our console for output regardless of redirect flags (horrible if we want to output our own process, but we can launch it in a separate command prompt to work around this kinda), and it forces us to push Enter at the end to terminate (horrible for "fully" automated processes, but we can work around it kinda by redirecting our own input to it), each of which we can work around independently, but the workarounds seem to be completely incompatible with one another.

As a secondary minor gripe, it does a lot of extra stuff we don't need.  It's hard to make a huge fuss over that, though, given how relatively inefficient this solution is for very large numbers of comparatively small files under in the name of "I/O is totes so much more expensive than CPU, yo", which I expect to not actually hold true very frequently here.

- P/Invoke libbsa is more complicated for me to figure out than this business, especially considering how this would have to change to work on 32-bit machines, if any (and I'm not willing to drop possible support for 32-bit machines just to avoid having to write 150 or so lines of code).

- Sharp.BSA.BA2 (well, BSA Browser) is licensed under GPLv3, and I don't really want to go to that level of restriction over, again, 150 or so lines of code.